requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
python-telegram-bot==22.3
httpx==0.28.1
matplotlib==3.10.3
pillow==11.3.0
pandas==2.3.1
scipy==1.14.1
pandas-ta==0.3.14b0
psycopg2-binary==2.9.9
redis==5.0.1
python-dotenv==1.0.0
